refactor: ♻️ consolidate all util functions into  app.core.util module
