+++ app/docs/api_docs/swagger_ui.py
@@ -6,6 +6,7 @@ from fastapi.openapi.docs import swagger_ui_default_parameters
 from starlette.responses import HTMLResponse
 
 from app.config.api_settings import get_settings
+from app.core.util import slugify
 from app.docs.api_docs.content.block import BLOCK_ENDPOINTS, UNICODE_BLOCK_OBJECT_INTRO, UNICODE_BLOCK_OBJECT_PROPERTIES
 from app.docs.api_docs.content.character import (
     CHARACTER_ENDPOINTS,
@@ -45,7 +46,6 @@ from app.docs.api_docs.content.intro import (
     SEARCH_HTML,
 )
 from app.docs.api_docs.content.plane import PLANE_ENDPOINTS, UNICODE_PLANE_OBJECT_INTRO, UNICODE_PLANE_OBJECT_PROPERTIES
-from app.docs.util import slugify
 
 
 # fmt: off
