+++ app/docs/dependencies/filter_tables.py
@@ -3,8 +3,8 @@ from sqlmodel import Session, distinct, select
 
 import app.db.models as db
 from app.core.cache import cached_data
+from app.core.util import slugify
 from app.db.engine import engine
-from app.docs.util import slugify
 from app.schemas.enums import (
     BidirectionalClass,
     CharacterFilterFlags,
